requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asn1crypto==0.24.0
astroid==2.4.2
attrs==19.3.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
click==7.1.2
cryptography==2.1.4
decorator==4.4.2
distlib==0.3.1
emoji==1.2.0
filelock==3.0.12
Flask==1.1.2
gunicorn==20.0.4
httplib2==0.9.2
idna==2.6
importlib-metadata==1.6.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
keyring==10.6.0
keyrings.alt==3.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
Mako==1.0.7
mccabe==0.6.1
more-itertools==8.2.0
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
packaging==20.3
pexpect==4.8.0
Pillow==5.1.0
pluggy==0.13.1
psutil==5.4.2
ptyprocess==0.6.0
py==1.8.1
pycairo==1.16.2
python-telegram-bot==13.1
pytimeparse==1.1.8
pytz==2020.4
pyxdg==0.25
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.13.2
six==1.14.0
systemd-python==234
timer3==0.1.2
toml==0.10.2
typed-ast==1.4.1
tzlocal==3.0b1
virtualenv==20.2.2
wadllib==1.3.2
wcwidth==0.1.9
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0

